fixed stdint.h

+++ librivac/include/stdint.h
@@ -26,12 +26,12 @@ typedef unsigned long long uint_least64_t;
 typedef signed char int_fast8_t;
 typedef int int_fast16_t;
 typedef int int_fast32_t;
-typedef long int_fast64_t;
+typedef long long int_fast64_t;
 
 typedef unsigned char uint_fast8_t;
 typedef unsigned int uint_fast16_t;
 typedef unsigned int uint_fast32_t;
-typedef unsigned long uint_fast64_t;
+typedef unsigned long long uint_fast64_t;
 
 // pointers
 typedef int intptr_t;
